Fixed NuGet configs so NuGet restores work again
The project file and packages.config were not consistent, and had pretty much broken the ability to restore the NuGet packages from nothing. Fixed the references, making packages.config consistent with newer references in the project file and converted a few PackageReference style references back to the old Reference style (the latter is what has to be used when using package.config, the two cannot be used in tandem). Also updated packages.config to reflect the .NET 4.8 migration.